참여자 추가 후 빈 배열 에러 해결
